Add Cloudfoundry support & deployment task to app
This commit alters slightly the application setup to support CloudFoundry deployment. It also adds a new file to represent a staging environment within the
app pipeline. We then have wrapped that in a make command to automate the deployment of said application to the staging environment.

This will be revisited when we add a production environment.